Integration tweaks
Rotates and sends a "v" for the victory toll at the end of the lap.
Added a button to tell the buggy to listen for the DTMF sound.
Now sends which direction a target box is approached from to the buggy.
Increased the area for displaying box values to allow for 3 values per
target box.
Removed reverse instructions from the buggy path.